feat: Create basic models to manage good of my site.
Product: represent common data for all goods.

Category: represent all allowed categories of goods

Attribute: represent a name of attibute and link the name to a paticulr category.
Each category has a set of name that is belongs to it.

ProductAttributeValue: represent a value of a paticular attribute for a paticular product.